+++ smallrye-reactive-messaging-pulsar/src/main/java/io/smallrye/reactive/messaging/pulsar/i18n/PulsarLogging.java
@@ -98,5 +98,5 @@ public interface PulsarLogging extends BasicLogger {
 
     @LogMessage(level = Logger.Level.WARN)
     @Message(id = 19025, value = "Failed to process a message sent to channel `%s`, tryign to reconsume in %s seconds, failure is: %s.")
-    void messageFailureDelayed(String channel, long delaySeconds , String reason);
+    void messageFailureDelayed(String channel, long delaySeconds, String reason);
 }
